Added final versions of happy-dynamic and happy_static
